Make it compile

+++ src/main/java/virtuoel/pehkui/mixin/EntitySelectorReaderMixin.java
@@ -24,9 +24,9 @@ public abstract class EntitySelectorReaderMixin implements PehkuiEntitySelectorR
 	abstract void setPredicate(Predicate<Entity> predicate);
 	
 	@Unique
-	NumberRange.FloatRange pehkui$scaleRange = NumberRange.FloatRange.ANY;
+	NumberRange.DoubleRange pehkui$scaleRange = NumberRange.DoubleRange.ANY;
 	@Unique
-	NumberRange.FloatRange pehkui$computedScaleRange = NumberRange.FloatRange.ANY;
+	NumberRange.DoubleRange pehkui$computedScaleRange = NumberRange.DoubleRange.ANY;
 	@Unique
 	ScaleType pehkui$scaleType = ScaleTypes.INVALID;
 	@Unique
@@ -61,13 +61,13 @@ public abstract class EntitySelectorReaderMixin implements PehkuiEntitySelectorR
 	}
 	
 	@Override
-	public NumberRange.FloatRange pehkui_getScaleRange()
+	public NumberRange.DoubleRange pehkui_getScaleRange()
 	{
 		return this.pehkui$scaleRange;
 	}
 	
 	@Override
-	public void pehkui_setScaleRange(final NumberRange.FloatRange baseScaleRange)
+	public void pehkui_setScaleRange(final NumberRange.DoubleRange baseScaleRange)
 	{
 		this.pehkui$scaleRange = baseScaleRange;
 	}
@@ -85,13 +85,13 @@ public abstract class EntitySelectorReaderMixin implements PehkuiEntitySelectorR
 	}
 	
 	@Override
-	public NumberRange.FloatRange pehkui_getComputedScaleRange()
+	public NumberRange.DoubleRange pehkui_getComputedScaleRange()
 	{
 		return this.pehkui$computedScaleRange;
 	}
 	
 	@Override
-	public void pehkui_setComputedScaleRange(final NumberRange.FloatRange computedScaleRange)
+	public void pehkui_setComputedScaleRange(final NumberRange.DoubleRange computedScaleRange)
 	{
 		this.pehkui$computedScaleRange = computedScaleRange;
 	}
